+++ CHANGELOG.md
@@ -1,10 +1,31 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [2.69.2] - 2019-02-20
+## [2.72.0] - 2019-02-20
 
 ### Added
 - `getPreviewFromPdf` method in `huh.utils.file`
+
+## [2.71.0] - 2019-02-20
+
+### Added
+- twig filter `image_data` in `ImageExtension` to get image data as array
+
+## [2.70.2] - 2019-02-19
+
+### Fixed
+- compile error due calling `System:getContainer` in `TemplateUtil` constructor
+
+## [2.70.1] - 2019-02-19
+
+### Fixed
+- `TemplateUtil::getTemplateGroup` now returns file extension in file name if not html.twig
+
+## [2.70.0] - 2019-02-18
+
+### Changed
+- `TemplateUtil` now supports all twig formats
+- `TemplateUtil::getTemplate` now throws an error if template not exist
 
 ## [2.69.1] - 2019-02-15